+++ packages/core/package.json
@@ -34,12 +34,25 @@
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/farm-fe"
+    "url": "https://github.com/farm-fe/farm"
   },
   "author": {
     "name": "bright wu",
     "email": "1521488775@qq.com"
   },
+  "keywords": [
+    "Farm",
+    "Vite",
+    "Build Tool",
+    "Rust Bundler",
+    "FarmFE",
+    "Farm Core",
+    "bundler",
+    "build-tool",
+    "frontend",
+    "dev-server",
+    "hmr"
+  ],
   "devDependencies": {
     "@napi-rs/cli": "^2.15.2",
     "@types/figlet": "^1.5.5",
@@ -84,6 +97,7 @@
     "@farmfe/runtime": "workspace:^0.9.2",
     "@farmfe/runtime-plugin-hmr": "workspace:^3.4.1",
     "@farmfe/runtime-plugin-import-meta": "workspace:*",
+    "@farmfe/utils": "workspace:*",
     "@koa/cors": "^4.0.0",
     "@swc/helpers": "^0.5.0",
     "browserslist-generator": "^2.1.0",
