+++ packages/create-farm-plugin/.eslintrc.json
@@ -1,0 +1,8 @@
+{
+  "root": true,
+  "extends": "../../.eslintrc.base.json",
+  "parserOptions": {
+    "project": ["./packages/create-farm-plugin/tsconfig.json"]
+  },
+  "rules": {}
+}
