+++ packages/create-farm-plugin/.gitignore
@@ -1,0 +1,1 @@
+build
