+++ packages/create-farm-plugin/README.md
@@ -1,15 +1,13 @@
-# create-farm
+# create-farm-plugin
 
 <div align="center">
   <a href="https://github.com/farm-fe/farm">
   <img src="../../assets/logo.png" width="550" />
   </a>
-  <p>Scaffolding Your First Farm Project</p>
+  <p>Scaffolding Your First Farm Plugin Project</p>
 </div>
 
-This package provides official templates for the Rust Web Bundler [Farm](https://github.com/farm-fe/farm).
-
-## Scaffolding Your First Farm Project
+## Scaffolding Your First Farm Plugin Project
 
 > **Compatibility Note:**
 > Farm requires [Node.js](https://nodejs.org/en/) version 16+. However, some templates require a higher Node.js version to work, please upgrade if your package manager warns about it.
@@ -17,19 +15,19 @@ This package provides official templates for the Rust Web Bundler [Farm](https:/
 With NPM:
 
 ```bash
-$ npm create farm@latest
+$ npm create farm-plugin@latest
 ```
 
 With Yarn:
 
 ```bash
-$ yarn create farm
+$ yarn create farm-plugin
 ```
 
 With PNPM:
 
 ```bash
-$ pnpm create farm
+$ pnpm create farm-plugin
 ```
 
 Then follow the prompts!
@@ -38,14 +36,14 @@ You can also directly specify the project name and the template you want to use 
 
 ```bash
 # npm 6.x
-npm create farm@latest my-vue-app --template react
+npm create farm@latest my-vue-app --template js
 
 # npm 7+, extra double-dash is needed:
-npm create farm@latest my-vue-app -- --template vue
+npm create farm@latest my-vue-app -- --template rust
 
 # yarn
-yarn create farm my-vue-app --template react
+yarn create farm my-vue-app --template js
 
 # pnpm
-pnpm create farm my-vue-app --template vue
+pnpm create farm my-vue-app --template rust
 ```
