+++ packages/create-farm-plugin/bin/create-farm-plugin.mjs
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+import '../dist/index.js';
