+++ packages/create-farm-plugin/farm.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from '@farmfe/core';
+
+export default defineConfig({
+  compilation: {
+    persistentCache: false,
+    output: {
+      targetEnv: 'node'
+    },
+    sourcemap: false,
+    presetEnv: false,
+  }
+});
