+++ packages/runtime-plugin-hmr/README.md
@@ -1,6 +1,8 @@
 # Note
 
-Farm is compatible with Vite. The Farm HMR client's API and implementation is similar to Vite's. see [License](./license).
+Runtime HMR Plugin for Rust Web Build Tool [Farm](https://github.com/farm-fe/farm).
+
+[Farm](https://github.com/farm-fe/farm) is compatible with Vite. The Farm HMR client's API and implementation is similar to Vite's. see [License](./license).
 
 - Farm's HMR client is compatible with Vite, see https://vitejs.dev/guide/api-hmr.html.
 - And it's inspired by both Vite and esm-hmr, see https://github.com/FredKSchott/esm-hmr
