+++ packages/runtime-plugin-hmr/package.json
@@ -4,9 +4,21 @@
   "description": "Runtime hmr plugin of Farm",
   "author": {
     "name": "bright wu",
-    "email": "1611331@mail.nankai.edu.cn"
+    "email": "1521488775@qq.com"
   },
   "main": "src/index.ts",
+  "keywords": [
+    "farm",
+    "farm runtime plugin",
+    "hmr",
+    "rust bundler",
+    "build tool",
+    "web bundler"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/farm-fe/farm"
+  },
   "scripts": {
     "build": "tsc -p tsconfig.json --noEmit",
     "type-check": "tsc -p tsconfig.json --noEmit"
