+++ packages/runtime-plugin-import-meta/README.md
@@ -1,0 +1,5 @@
+# Note
+
+Runtime Meta Plugin for Rust Web Build Tool [Farm](https://github.com/farm-fe/farm).
+
+Support and polyfill `import.meta.env` for Farm Runtime.
