+++ packages/runtime-plugin-import-meta/package.json
@@ -7,6 +7,17 @@
     "email": "q502489978@live.com"
   },
   "main": "src/index.ts",
+  "keywords": [
+    "farm",
+    "farm runtime plugin",
+    "rust bundler",
+    "build tool",
+    "web bundler"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/farm-fe/farm"
+  },
   "scripts": {
     "build": "tsc -p tsconfig.json --noEmit",
     "type-check": "tsc -p tsconfig.json --noEmit"
