+++ packages/runtime/README.md
@@ -1,0 +1,3 @@
+# Note
+
+Runtime support for Rust Web Build Tool [Farm](https://github.com/farm-fe/farm).
