+++ packages/utils/.eslintrc.json
@@ -1,0 +1,8 @@
+{
+  "root": true,
+  "extends": "../../.eslintrc.base.json",
+  "parserOptions": {
+    "project": ["./packages/utils/tsconfig.json"]
+  },
+  "rules": {}
+}
