+++ packages/utils/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "@farmfe/utils",
+  "version": "0.0.0",
+  "description": "farmfe utils",
+  "main": "dist/index.js",
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "default": "./dist/index.js"
+    },
+    "./colors": {
+      "import": "./dist/color.js",
+      "default": "./dist/color.js"
+    }
+  },
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/farm-fe/farm"
+  },
+  "author": {
+    "name": "bright wu",
+    "email": "1521488775@qq.com"
+  },
+  "keywords": [
+    "Farm",
+    "Vite",
+    "Build Tool",
+    "Rust Bundler",
+    "FarmFE",
+    "Farm Core",
+    "bundler",
+    "build-tool",
+    "frontend",
+    "dev-server",
+    "hmr"
+  ],
+  "files": [
+    "dist"
+  ]
+}
